requirements.txt
discord.py
PyNaCl
aiohttp
boticordpy
numexpr
motor
fluent.runtime
jishaku
python-dotenv
distro
asyncstdlib